BAVL-1295 removing old email code now the refactored email code is active in production.
