+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/config/FeatureSwitches.kt
@@ -31,7 +31,6 @@ class FeatureSwitches(private val environment: Environment) {
 
 enum class BooleanFeature(val label: String) {
   FEATURE_PLACEHOLDER("feature.placeholder.example"),
-  FEATURE_EMAIL_FACADE_ENABLED("feature.email.facade.enabled"),
 }
 
 enum class StringFeature(val label: String) {
