--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/service/OriginalBookingFacade.kt
@@ -1,228 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service
-
-import org.slf4j.LoggerFactory
-import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.client.prisonersearch.PrisonerSearchClient
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.config.EmailService
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.config.VideoBookingEmail
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.BookingContact
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.BookingType.COURT
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.BookingType.PROBATION
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.ContactType
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.Notification
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.StatusCode
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.entity.VideoBooking
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.Prisoner
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.request.AmendVideoBookingRequest
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.request.CreateVideoBookingRequest
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.request.RequestVideoBookingRequest
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.repository.AdditionalBookingDetailRepository
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.repository.NotificationRepository
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.repository.PrisonRepository
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.emails.court.CourtEmailFactory
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.emails.probation.ProbationEmailFactory
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.events.DomainEventType
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.events.OutboundEventsService
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.locations.LocationsService
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.locations.availability.AvailabilityService
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.telemetry.TelemetryEventFactory
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.telemetry.TelemetryService
-import java.time.LocalDate
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.model.request.BookingType as RequestBookingType
-
-/**
- * This facade exists to ensure all booking related transactions are fully committed prior to sending any events or emails.
- * That is to say, ideally calls to the facade should not be wrapped in transactions.
- */
-@Component
-class OriginalBookingFacade(
-  private val videoBookingServiceDelegate: VideoBookingServiceDelegate,
-  private val contactsService: ContactsService,
-  private val prisonRepository: PrisonRepository,
-  private val emailService: EmailService,
-  private val notificationRepository: NotificationRepository,
-  private val outboundEventsService: OutboundEventsService,
-  private val locationsService: LocationsService,
-  private val prisonerSearchClient: PrisonerSearchClient,
-  private val telemetryService: TelemetryService,
-  private val availabilityService: AvailabilityService,
-  private val additionalBookingDetailRepository: AdditionalBookingDetailRepository,
-  private val changeTrackingService: ChangeTrackingService,
-) : BookingFacade {
-  companion object {
-    private val log = LoggerFactory.getLogger(this::class.java)
-  }
-
-  override fun create(bookingRequest: CreateVideoBookingRequest, createdBy: User): Long {
-    require(createdBy is PrisonUser || availabilityService.isAvailable(bookingRequest)) {
-      if (bookingRequest.bookingType == RequestBookingType.COURT) {
-        "Unable to create court booking, booking overlaps with an existing appointment."
-      } else {
-        "Unable to create probation booking, booking overlaps with an existing appointment."
-      }
-    }
-
-    val (booking, prisoner) = videoBookingServiceDelegate.create(bookingRequest, createdBy)
-    outboundEventsService.send(DomainEventType.VIDEO_BOOKING_CREATED, booking.videoBookingId)
-    sendEmails(BookingAction.CREATE, booking, prisoner, createdBy)
-    trackTelemetry(BookingAction.CREATE, booking, createdBy)
-    return booking.videoBookingId
-  }
-
-  override fun amend(videoBookingId: Long, bookingRequest: AmendVideoBookingRequest, amendedBy: User): Long {
-    require(amendedBy is PrisonUser || availabilityService.isAvailable(videoBookingId, bookingRequest)) {
-      if (bookingRequest.bookingType == RequestBookingType.COURT) {
-        "Unable to amend court booking, booking overlaps with an existing appointment."
-      } else {
-        "Unable to amend probation booking, booking overlaps with an existing appointment."
-      }
-    }
-
-    // Need to check before amendment to see before picture.
-    val changeType = changeTrackingService.determineChangeType(videoBookingId, bookingRequest, amendedBy)
-
-    // Amend regardless of changes (this is in-case new fields ever are introduced but not covered by the change check).
-    val (booking, prisoner) = videoBookingServiceDelegate.amend(videoBookingId, bookingRequest, amendedBy)
-    outboundEventsService.send(DomainEventType.VIDEO_BOOKING_AMENDED, videoBookingId)
-
-    // Only send emails on back of change check above.
-    if (changeType != ChangeType.NONE) {
-      sendEmails(BookingAction.AMEND, booking, prisoner, amendedBy, changeType)
-    } else {
-      log.info("No changes detected for video booking $videoBookingId, not sending email")
-    }
-
-    trackTelemetry(BookingAction.AMEND, booking, amendedBy)
-
-    return videoBookingId
-  }
-
-  override fun cancel(videoBookingId: Long, cancelledBy: PrisonUser) {
-    cancelBooking(videoBookingId, cancelledBy)
-  }
-
-  override fun cancel(videoBookingId: Long, cancelledBy: ExternalUser) {
-    cancelBooking(videoBookingId, cancelledBy)
-  }
-
-  override fun request(booking: RequestVideoBookingRequest, user: ExternalUser) {
-    videoBookingServiceDelegate.request(booking, user)
-  }
-
-  private fun cancelBooking(videoBookingId: Long, cancelledBy: User) {
-    val booking = videoBookingServiceDelegate.cancel(videoBookingId, cancelledBy)
-    log.info("Video booking ${booking.videoBookingId} cancelled by user type ${cancelledBy::class.simpleName}")
-    outboundEventsService.send(DomainEventType.VIDEO_BOOKING_CANCELLED, videoBookingId)
-    sendEmails(BookingAction.CANCEL, booking, getPrisoner(booking.prisoner()), cancelledBy)
-    trackTelemetry(BookingAction.CANCEL, booking, cancelledBy)
-  }
-
-  override fun cancel(videoBookingId: Long, cancelledBy: ServiceUser) {
-    // No events should be raised on the back of calling this function, this is by design.
-    val booking = videoBookingServiceDelegate.cancel(videoBookingId, cancelledBy)
-    log.info("Video booking ${booking.videoBookingId} cancelled by user type ${cancelledBy::class.simpleName}")
-    sendEmails(BookingAction.CANCEL, booking, getPrisoner(booking.prisoner()), cancelledBy)
-    trackTelemetry(BookingAction.CANCEL, booking, cancelledBy)
-  }
-
-  override fun courtHearingLinkReminder(videoBooking: VideoBooking, user: ServiceUser) {
-    require(videoBooking.isBookingType(COURT) && videoBooking.isStatus(StatusCode.ACTIVE)) { "Video booking with id ${videoBooking.videoBookingId} must be an active court booking" }
-    require(videoBooking.court!!.enabled) { "Video booking with id ${videoBooking.videoBookingId} is not with an enabled court" }
-    require(videoBooking.appointments().any { it.appointmentDate > LocalDate.now() }) { "Video booking with id ${videoBooking.videoBookingId} must be after today" }
-    require(videoBooking.videoUrl == null) { "Video booking with id ${videoBooking.videoBookingId} already has a court hearing link" }
-    sendEmails(BookingAction.COURT_HEARING_LINK_REMINDER, videoBooking, getPrisoner(videoBooking.prisoner()), user)
-  }
-
-  override fun sendProbationOfficerDetailsReminder(videoBooking: VideoBooking, user: ServiceUser) {
-    require(videoBooking.isBookingType(PROBATION) && videoBooking.isStatus(StatusCode.ACTIVE)) { "Video booking with id ${videoBooking.videoBookingId} must be an active probation booking" }
-    require(videoBooking.probationTeam!!.enabled) { "Video booking with id ${videoBooking.videoBookingId} is not with an enabled probation team" }
-    require(videoBooking.appointments().any { it.appointmentDate > LocalDate.now() }) { "Video booking with id ${videoBooking.videoBookingId} must be after today" }
-    sendProbationEmails(BookingAction.PROBATION_OFFICER_DETAILS_REMINDER, videoBooking, getPrisoner(videoBooking.prisoner()), user)
-  }
-
-  override fun prisonerTransferred(videoBookingId: Long, user: ServiceUser) {
-    val booking = videoBookingServiceDelegate.cancel(videoBookingId, user)
-    log.info("Video booking ${booking.videoBookingId} cancelled due to transfer")
-    outboundEventsService.send(DomainEventType.VIDEO_BOOKING_CANCELLED, videoBookingId)
-    sendEmails(BookingAction.TRANSFERRED, booking, getReleasedOrTransferredPrisoner(booking.prisoner()), user)
-    trackTelemetry(BookingAction.TRANSFERRED, booking, user)
-  }
-
-  override fun prisonerReleased(videoBookingId: Long, user: ServiceUser) {
-    val booking = videoBookingServiceDelegate.cancel(videoBookingId, user)
-    log.info("Video booking ${booking.videoBookingId} cancelled due to release")
-    outboundEventsService.send(DomainEventType.VIDEO_BOOKING_CANCELLED, videoBookingId)
-    sendEmails(BookingAction.RELEASED, booking, getReleasedOrTransferredPrisoner(booking.prisoner()), user)
-    trackTelemetry(BookingAction.RELEASED, booking, user)
-  }
-
-  private fun trackTelemetry(action: BookingAction, booking: VideoBooking, user: User) {
-    TelemetryEventFactory.event(action, booking, user)?.let(telemetryService::track)
-  }
-
-  private fun getPrisoner(prisonerNumber: String) = prisonerSearchClient.getPrisoner(prisonerNumber)!!.let { Prisoner(it.prisonerNumber, it.prisonId!!, it.firstName, it.lastName, it.dateOfBirth) }
-
-  private fun getReleasedOrTransferredPrisoner(prisonerNumber: String) = prisonerSearchClient.getPrisoner(prisonerNumber)!!.let { Prisoner(it.prisonerNumber, it.lastPrisonId!!, it.firstName, it.lastName, it.dateOfBirth) }
-
-  private fun sendEmails(action: BookingAction, booking: VideoBooking, prisoner: Prisoner, user: User, changeType: ChangeType = ChangeType.GLOBAL) {
-    when (booking.bookingType) {
-      COURT -> sendCourtEmails(action, booking, prisoner, user, changeType)
-      PROBATION -> sendProbationEmails(action, booking, prisoner, user, changeType)
-    }
-  }
-
-  private fun sendCourtEmails(eventType: BookingAction, booking: VideoBooking, prisoner: Prisoner, user: User, changeType: ChangeType = ChangeType.GLOBAL) {
-    val (pre, main, post) = Triple(booking.preHearing(), booking.mainHearing()!!, booking.postHearing())
-    val prison = prisonRepository.findByCode(booking.prisonCode())!!
-    val contacts = contactsService.getBookingContacts(booking.videoBookingId, user).withAnEmailAddress()
-    val locations = setOfNotNull(pre?.prisonLocationId, main.prisonLocationId, post?.prisonLocationId).mapNotNull { locationsService.getLocationById(it) }.associateBy { it.dpsLocationId }
-
-    val emails = contacts.mapNotNull { contact ->
-      when (contact.contactType) {
-        ContactType.USER -> CourtEmailFactory.user(contact, prisoner, booking, prison, main, pre, post, locations, eventType).takeIf { user is PrisonUser || user is ExternalUser }
-        ContactType.COURT -> CourtEmailFactory.court(contact, prisoner, booking, prison, main, pre, post, locations, eventType).takeIf { (user is PrisonUser || user is ServiceUser) && changeType in setOf(ChangeType.GLOBAL) }
-        ContactType.PRISON -> CourtEmailFactory.prison(contact, prisoner, booking, prison, contacts, main, pre, post, locations, eventType).takeIf { changeType in setOf(ChangeType.GLOBAL, ChangeType.PRISON) }
-        else -> null
-      }
-    }
-
-    emails.forEach { courtEmail -> sendEmailAndSaveNotification(courtEmail, booking, eventType) }
-  }
-
-  private fun sendProbationEmails(eventType: BookingAction, booking: VideoBooking, prisoner: Prisoner, user: User, changeType: ChangeType = ChangeType.GLOBAL) {
-    val appointment = booking.appointments().single()
-    val prison = prisonRepository.findByCode(booking.prisonCode())!!
-    val contacts = contactsService.getBookingContacts(booking.videoBookingId, user).withAnEmailAddress()
-    val location = locationsService.getLocationById(appointment.prisonLocationId)!!
-    val additionalBookingDetail = additionalBookingDetailRepository.findByVideoBooking(booking)
-
-    val emails = contacts.mapNotNull { contact ->
-      when (contact.contactType) {
-        ContactType.USER -> ProbationEmailFactory.user(contact, prisoner, booking, prison, appointment, location, eventType, additionalBookingDetail).takeIf { user is PrisonUser || user is ExternalUser }
-        ContactType.PROBATION -> ProbationEmailFactory.probation(contact, prisoner, booking, prison, appointment, location, eventType, additionalBookingDetail).takeIf { (user is PrisonUser || user is ServiceUser) && changeType in setOf(ChangeType.GLOBAL) }
-        ContactType.PRISON -> ProbationEmailFactory.prison(contact, prisoner, booking, prison, appointment, location, eventType, contacts, additionalBookingDetail).takeIf { changeType in setOf(ChangeType.GLOBAL, ChangeType.PRISON) }
-        else -> null
-      }
-    }
-
-    emails.forEach { probationEmail -> sendEmailAndSaveNotification(probationEmail, booking, eventType) }
-  }
-
-  private fun sendEmailAndSaveNotification(email: VideoBookingEmail, booking: VideoBooking, action: BookingAction) {
-    emailService.send(email).onSuccess { (govNotifyId, templateId) ->
-      notificationRepository.saveAndFlush(
-        Notification(
-          videoBooking = booking,
-          email = email.address,
-          govNotifyNotificationId = govNotifyId,
-          templateName = templateId,
-          reason = action.name,
-        ),
-      )
-    }.onFailure {
-      log.info("BOOKINGS: Failed to send ${action.name} email for video booking ID ${booking.videoBookingId}.")
-    }
-  }
-
-  private fun Collection<BookingContact>.withAnEmailAddress() = filter { it.email != null }
-}
