+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/integration/config/FeatureSwitchesTest.kt
@@ -31,7 +31,6 @@ class FeatureSwitchesTest : IntegrationTestBase() {
       "feature.public.private.comments.sync=true",
       "feature.court.only.prisons=PVI",
       "feature.probation.only.prisons=MDI",
-      "feature.email.facade.enabled=true",
     ],
   )
   @Nested
